Address code review: BigDecimal for amounts, saveAll() batch, deprecation comment
Agent-Logs-Url: https://github.com/ajarvis3/rate-limit-api/sessions/45a7dc64-ae1c-4497-bfaa-6520ffd00c8c

+++ dunning-service/src/main/java/com/ratelimit/dunning/model/DunningRecord.java
@@ -1,6 +1,7 @@
 package com.ratelimit.dunning.model;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "dunning_record")
@@ -12,14 +13,14 @@ public class DunningRecord {
 
     private String userId;
     private Long invoiceId;
-    private Double amount;
+    private BigDecimal amount;
     private long failedAt;
     private long retryAfter;
     private boolean retried;
 
     public DunningRecord() {}
 
-    public DunningRecord(String userId, Long invoiceId, Double amount, long failedAt, long retryAfter) {
+    public DunningRecord(String userId, Long invoiceId, BigDecimal amount, long failedAt, long retryAfter) {
         this.userId = userId;
         this.invoiceId = invoiceId;
         this.amount = amount;
@@ -36,8 +37,8 @@ public class DunningRecord {
     public Long getInvoiceId() { return invoiceId; }
     public void setInvoiceId(Long invoiceId) { this.invoiceId = invoiceId; }
 
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 
     public long getFailedAt() { return failedAt; }
     public void setFailedAt(long failedAt) { this.failedAt = failedAt; }
